Moved Header.js and set prettier to work with jsx files too
